two_level_morphanalyzer/requirements.txt
asgiref==3.6.0
beautifulsoup4==4.11.2
bs4==0.0.1
click==8.1.3
colorama==0.4.6
Django==4.1.7
django-ckeditor==6.5.1
django-debug-toolbar==4.2.0
django-extensions==3.2.1
django-js-asset==2.0.0
django-ranged-response==0.2.0
django-recaptcha==3.0.0
django-simple-captcha==0.5.20
djangorestframework==3.14.0
et-xmlfile==1.1.0
importlib-metadata==6.0.0
joblib==1.2.0
Markdown==3.4.3
openpyxl==3.1.2
Pillow==9.4.0
pydub==0.25.1
pytz==2022.7.1
regex==2023.3.23
soupsieve==2.4
sqlparse==0.4.3
tqdm==4.65.0
tzdata==2023.2
zipp==3.15.0
